test(harness): move live Azure chat tests to chat_live_azure suite
Per audit 5/8c:
- test_azure_ai.py: 7 live tests (request_format, gpt5 reasoning,
  completion_azure, flexible api_base, model-router x3) moved with the
  AzureModelRouterStreamingCallback helper they alone used; the
  map_azure_model_group / image-url body / deepseek reasoning goldens stay
- test_azure_openai.py: test_azure_safety_result,
  test_completion_azure_deployment_id, test_azure_openai_with_prompt_cache_key
  moved; header/url-builder/param-mapping goldens stay; TestAzureEmbedding
  stays (non-chat, Sameer)
- test_azure_o_series.py: TestAzureOpenAIO3Mini moved; its keep-unit
  test_override_fake_stream stays behind as
  test_azure_o1_override_fake_stream (no network; router/model-info golden)
